renamed functionDetector.js to codeParser.js. Refactored for readability. Function is now detected when the cursor is in the function definition. Began working on logic to display error if the cursor is not in a function and generate docstring otherwise.
